Typographic conventions proposition.
Code conventions are good. Typographic conventions are too!

I'm proposing a few changes in this document to make it more compliant to French (and more general) standards.

* Use proper capitalization for software names (JavaScript instead of javascript, HTML instead of html, same for XML, GitHub, Git, Less, Sass and Stylus).
* Always use a non-breaking space before a colon in French.
* No comma before ellipsis.
* English typography in fully English titles.
* Proper case and punctuation for lists (either a sentence splitted in a list with commas, or a sentence per list item).
* Space after `function` keyword even if anonymous, that's my personal touch in these coding conventions. :3

Also, it looks like there's some inconsistencies between what to put in italic or bold... usually non-French words, software names and acronyms would be in italic, but I haven't heard of anything strict in this subject.

I want to make it clear, this is only a suggestion because I care about these details. Be sure to review the changes, and only include (some of) them if you agree with these conventions.

BTW, I think it's wrong to force students to use GitHub. Maybe it's for practical reasons, but while I'm totally in favor of using Git on whatever server they want, I'm against letting no other choice than using a proprietary and centralized service to host their work. Maybe it can be a good idea to evoke self-hosting, which is really simple with a Raspberry PI or equivalent and a little configuration for port forwarding, but I know this is totally out of scope for this commit.